+++ src/client/widgets/sliders/range.js
@@ -214,7 +214,8 @@ class Range extends Fader {
             d2 = Math.round(fader.percentToCoord(fader.valueToPercent(this.faders[this.getProp('horizontal')?1:0].value))),
             m = this.getProp('horizontal') ? this.height / 2 : this.width / 2,
             dashed = this.getProp('dashed'),
-            compact = this.getProp('design') === 'compact'
+            compact = this.getProp('design') === 'compact',
+            knobHeight = this.cssVars.knobSize * PXSCALE, knobWidth = knobHeight * .6 * PXSCALE
 
         this.clear()
 
@@ -306,7 +307,7 @@ class Range extends Fader {
                     this.ctx.fillStyle = this.cssVars.colorBg
 
                     this.ctx.beginPath()
-                    this.ctx.rect(m - 6 * PXSCALE, _d - 10 * PXSCALE, 12 * PXSCALE, 20 * PXSCALE)
+                    this.ctx.rect(m - knobWidth / 2, _d - knobHeight / 2, knobWidth, knobHeight)
                     this.ctx.fill()
 
                     this.ctx.globalAlpha = this.cssVars.alphaStroke
@@ -314,7 +315,7 @@ class Range extends Fader {
                     this.ctx.lineWidth = PXSCALE
 
                     this.ctx.beginPath()
-                    this.ctx.rect(m - 5.5 * PXSCALE, _d - 9.5 * PXSCALE, 11 * PXSCALE, 19 * PXSCALE)
+                    this.ctx.rect(m - knobWidth / 2 + 0.5 * PXSCALE, _d - knobHeight / 2  + 0.5 * PXSCALE, knobWidth - PXSCALE, knobHeight - PXSCALE)
                     this.ctx.stroke()
 
                 }
@@ -328,7 +329,7 @@ class Range extends Fader {
 
             }
 
-            this.clearRect = [m - 10 * PXSCALE, this.gaugePadding - 10 * PXSCALE, 20 * PXSCALE, height - 2 * this.gaugePadding + 20 * PXSCALE]
+            this.clearRect = [m - knobWidth / 2 - PXSCALE, this.gaugePadding - knobHeight / 2  - PXSCALE, knobWidth + 2 * PXSCALE, height - 2 * this.gaugePadding + knobHeight + 2 * PXSCALE]
 
         } else {
 
@@ -366,7 +367,7 @@ class Range extends Fader {
         if (this.getProp('pips')) {
             this.ctx.globalAlpha = 1
             this.ctx.drawImage(this.pips, 0, 0)
-            if (!compact) this.clearRect = [this.clearRect, [m + 10 * PXSCALE, 0, 10 * PXSCALE + this.pipsTextSize, height]]
+            if (!compact) this.clearRect = [this.clearRect, [m + knobWidth / 2, 0, 12 * PXSCALE + this.pipsTextSize, height]]
         }